requirements.txt
num2words==0.5.12
numpy==1.23.5
pandas==1.5.2
PythonTurtle==0.3.2
Pillow==9.3.0

